cc: Extract configuration tabs' order to a separate file and modify how the order is fetched
